# APIDocumentation.yml
openapi: 3.0.3
info:
  title: Attendance Service API
  description: |-
    This is the API documentation of Attendance Service (MOSIP).
  version: 1.0.0
servers:
  - url: http://20.235.163.35:8080
...
paths:
...
  /admin/lecture/create:
    post:
      tags:
        - Lectures
      summary: Create a lecture
      description: This endpoint will provide functionality for a admin to create a lecture.
      operationId: create
      requestBody:
        description: Lecture details
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateLectureRequestDTO'
        required: true
      responses:
        '200':
          description: Lecture created successfully.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "LECTURE_CREATED_SUCCESSFULLY"
                  message:
                    type: string
                    example: "Lecture created successfully!"
                  data:
                    type: number
                    description: Lecture id
                    example: 2
        '400':
          description: Invalid or missing details
      security:
        - access_token: [ ]
  /admin/lecture/start-lecture/{lectureId}:
    get:
      tags:
        - Lectures
      summary: Start a lecture
      description: This endpoint will provide functionality for a admin to start a lecture.
      operationId: startLecture
      parameters:
        - name: lectureId
          in: path
          description: ID of the lecture
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lecture started successfully.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "Lecture started successfully"
                  data:
                    type: integer
                    format: int64
                    description: "Lecture id"
                    example: 43872
        '400':
          description: Invalid or missing details
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "INVALID_DATA"
                  message:
                    type: string
                    example: "Lecture already started"
      security:
        - access_token: [ ]
  /admin/lecture/end-lecture/{lectureId}:
    get:
      tags:
        - Lectures
      summary: End a lecture
      description: This endpoint will provide functionality for a admin to end a lecture.
      operationId: endLecture
      parameters:
        - name: lectureId
          in: path
          description: ID of the lecture
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lecture ended successfully.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "Lecture finished successfully"
                  data:
                    type: integer
                    format: int64
                    description: "Lecture id"
                    example: 43872
        '400':
          description: Invalid or missing details
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "INVALID_DATA"
                  message:
                    type: string
                    example: "Invalid lecture"
      security:
        - access_token: []
  /admin/lecture/getAllLectures:
    get:
      tags:
        - Lectures
      summary: Get all lectures
      description: This endpoint will provide details of all lectures of the system.
      operationId: getAllLectures
      responses:
        '200':
          description: Lectures fetched successfully.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "Lectures Fetched."
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LectureDTO'
        '404':
          description: Lectures not found.
      security:
        - access_token: [ ]
  /admin/lecture/get-lecture/{lectureId}:
    get:
      tags:
        - Lectures
      summary: Get a lecture by ID
      description: This endpoint will provide details of a lecture for a given exam ID.
      operationId: getLectureById
      parameters:
        - name: lectureId
          in: path
          description: ID of the Lecture
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Lecture fetched successfully.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "Lecture Fetched."
                  data:
                    $ref: '#/components/schemas/LectureDTO'
        '400':
          description: Invalid ID supplied or ID missing
        '404':
          description: Lecture ID is not found.
      security:
        - access_token: [ ]
  /admin/lecture/getCurrentLectures:
    get:
      tags:
        - Lectures
      summary: Get currently started lectures
      description: This endpoint will provide functionality for a lecturer to view his/her lectures that are currently started.
      operationId: getCurrentLectures
      responses:
        '200':
          description: Lectures Fetched.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "Lectures Fetched"
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/LectureDTO'

      security:
        - access_token: [ ]
  /student/lecture-attendance/mark-lecture-attendance:
    post:
      tags:
        - Lectures
      summary: Mark attendance for a lecture
      description: This endpoint will provide functionality for a student to mark their attendance in a lecture and display their details.
      operationId: markLectureAttendance
      requestBody:
        description: Biometric credentials and lecture id
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MarkAttendanceRequestDTO'
        required: true
      responses:
        '200':
          description: Attendance marked successfully.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "OK"
                  message:
                    type: string
                    example: "Attendance marked successfully"
                  data:
                    $ref: '#/components/schemas/StudentDTO'
        '400':
          description: Invalid or missing details
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "INVALID_DATA"
                  message:
                    type: string
                    example: "Attendance already marked"
      security:
        - access_token: [ ]
  /admin/lecture/deleteLecture/{lectureId}:
    delete:
      tags:
        - Lectures
      summary: Delete a lecture
      description: This endpoint will provide functionality for a admin to delete a lecture.
      operationId: deleteLectureByID
      parameters:
        - name: lectureId
          in: path
          description: ID of the lecture
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Deleted lecture successfully.
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "LECTURE_DELETED_SUCCESSFULLY"
                  message:
                    type: string
                    example: "Deleted lecture successfully!"
                  data:
                    type: integer
                    format: int64
                    description: "Lecture id"
                    example: 43872
        '400':
          description: Invalid or missing details
          content:
            application/json:
              schema:
                properties:
                  status:
                    type: string
                    example: "LECTURE_NOT_FOUND"
                  message:
                    type: string
                    example: "No Lecture found corresponding to the lecture ID!"
      security:
        - access_token: [ ]
...
components:
  schemas:
...
    CreateLectureRequestDTO:
      type: object
      properties:
        course_id:
          type: integer
          example: 65
        hall_id:
          type: integer
          example: 11
        lecturer_id:
          type: integer
          example: 37
        start_time:
          type: string
          format: date-time
          example: "2023-07-15T10:00:00"
        end_time:
          type: string
          format: date-time
          example: "2023-07-15T12:00:00"
        expected_attendance:
          type: integer
          example: 128
...
    Fingerprint:
      type: object
      properties:
        buffer:
          type: object
          properties:
            type:
              type: string
              example: Buffer
            data:
              type: array
              example: [255,216,255,224,0,16,74,70,73,70,0,1,1,0,0,1,0,1,0,0,255,254,0,32,67,114,101,97,116,101,100,32,98,121,32,73,110,169,18,75,52,142,169,110,62,240,216,163,1,190,189,43,207,126,58,248,182,120,181,11,63,15,233,215,206,187,131,61,226,138,0,41,13,20,80,2,82,209,69,0,45,20,81,64,5,20,81,64,5,20,81,64,5,20,81,64,5,20,81,64,5,20,81,64,5,20,81,64,31,255,217]
        bioSubType:
          type: string
          example: Right Thumb
...
    LectureDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: ID of lecture
          example: 32
          nullable: false
        attendance:
          type: integer
          format: int64
          description: Number of students attending
          example: 32
          nullable: false
        start_time:
          type: string
          format: date-time
          example: 2023-07-15T11:00:00
        end_time:
          type: string
          format: date-time
          example: 2023-07-15T13:00:00
        expected_attendance:
          type: integer
          example: 128
        moduleCode:
          type: string
          example: "CS1012"
        moduleName:
          type: string
          example: "Programming Fundamentals"
        intake:
          type: integer
          description: Intake of course
          example: 22
        hall_id:
          type: integer
          example: 6
        lecturer_id:
          type: integer
          example: 6
        course_id:
          type: integer
          example: 6
        is_started:
          type: boolean
          example: true
        is_ended:
          type: boolean
          example: false
...
    MarkAttendanceRequestDTO:
      type: object
      properties:
        fingerprint:
          $ref: '#/components/schemas/Fingerprint'
        eventId:
          type: integer
          format: int64
          description: ID of lecture or exam
          example: 6513213
          nullable: false
...
    StudentDTO:
      type: object
      properties:
        index_number:
          type: string
          description: University index number
          example: 190429G
          nullable: false
        id:
          type: string
          description: MOSIP id
          example: 99
          nullable: false
        name:
          type: string
          description: Name of student
          example: D.T.A.W.S.Paranayapa
          nullable: false
        picture:
          type: string
          format: base64
          description: Picture of student
          example: 23xcewjf4wyfcw4093ixh32feefr4f4cgttrbtrbt
          nullable: false
...
  securitySchemes:
    access_token:
      type: apiKey
      name: Access-Token
      in: header